move hooks to MakeOfferModal/hooks directory
